// Exercises 9.8-9.9-9.21-9.30/patientor/src/types.ts
export interface Diagnosis {
  code: string;
  name: string;
  latin?: string;
}

export enum Gender {
  Male = "male",
  Female = "female",
  Other = "other"
}

export interface Patient {
  id: string;
  name: string;
  occupation: string;
  gender: Gender;
  ssn?: string;
  dateOfBirth?: string;
  entries: Entry[]
}

export type PatientInfo = Omit<Patient, 'id'>;

export type PatientFormValues = Omit<Patient, "id" | "entries">;

// Base Entry
interface BaseEntry {
  id: string;
  description: string;
  date: string;
  specialist: string;
  diagnosisCodes?: Array<Diagnosis['code']>;
}

export enum HealthCheckRating {
  "Healthy" = 0,
  "LowRisk" = 1,
  "HighRisk" = 2,
  "CriticalRisk" = 3
}

// Health Check Entry
export  interface HealthCheckEntry extends BaseEntry {
  type: "HealthCheck";
  healthCheckRating: HealthCheckRating;
}

// Hospital Entry
export  interface HospitalEntry extends BaseEntry {
  type: "Hospital";
  discharge: {
    date: string;
    criteria: string;
  };
}

// Occupational Healthcare Entry
export interface OccupationalHealthcareEntry extends BaseEntry {
  type: "OccupationalHealthcare";
  employerName: string;
  sickLeave?: {
    startDate: string;
    endDate: string;
  };
}

// Combined Entry type
export type Entry = 
  | HospitalEntry 
  | OccupationalHealthcareEntry 
  | HealthCheckEntry;


// First, define the UnionOmit utility type
type UnionOmit<T, K extends string | number | symbol> = 
  T extends unknown ? Omit<T, K> : never;

export  type EntryWithoutId = UnionOmit<Entry, 'id'>; 